requirements.txt
customtkinter==5.2.2
easyocr==1.7.2
paddleocr==3.4.0
opencv-python==4.13.0.92
Pillow==11.3.0
numpy==2.4.2
google-genai
simple-lama-inpainting
